bugged

+++ database/migrations/2023_08_31_035901_create_profiles_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('profiles', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('user_id');
+            $table->integer('user_id');
             $table->string('title')->nullable();
             $table->text('description')->nullable();
             $table->string('url')->nullable();
